+++ services/sso/src/__tests__/dashboard/users/getMe.test.ts
@@ -9,10 +9,17 @@ import { dashboardLoginAndGetToken } from '../../helpers/testHelpers'
 
 const route = `${pathRoot.v1.dashboard.users}/me`
 let adminAuthToken = ''
+let mentorAuthToken = ''
+
 beforeAll(async () => {
   adminAuthToken = await dashboardLoginAndGetToken(
     testUserData.admin.dni,
     testUserData.admin.password
+  )
+
+  mentorAuthToken = await dashboardLoginAndGetToken(
+    testUserData.mentor.dni,
+    testUserData.mentor.password
   )
 })
 
@@ -26,7 +33,7 @@ afterAll(async () => {
 })
 
 describe('Testing get users/me endpoint', () => {
-  it('should succeed with a valid token', async () => {
+  it('should succeed with an admin valid token', async () => {
     const response = await supertest(server)
       .get(route)
       .set('Cookie', [adminAuthToken])
@@ -34,11 +41,25 @@ describe('Testing get users/me endpoint', () => {
     expect(response.body.dni).toBeTypeOf('string')
     expect(
       userSchema
-        .pick({ dni: true, email: true, name: true })
+        .pick({ dni: true, email: true, name: true, role: true })
+        .safeParse(response.body).success
+    ).toBeTruthy()
+  })
+
+  it('should succeed with a mentor valid token', async () => {
+    const response = await supertest(server)
+      .get(route)
+      .set('Cookie', [mentorAuthToken])
+    expect(response.status).toBe(200)
+    expect(response.body.dni).toBeTypeOf('string')
+    expect(
+      userSchema
+        .pick({ dni: true, email: true, name: true, role: true })
         .safeParse(response.body).success
     ).toBeTruthy()
   })
 })
+
 describe('User Status Handling in Get users/me Endpoint', () => {
   afterEach(async () => {
     await client.query('UPDATE "user" SET status = $1 WHERE dni = $2', [
@@ -46,6 +67,7 @@ describe('User Status Handling in Get users/me Endpoint', () => {
       testUserData.admin.dni,
     ])
   })
+
   it('should fail when the logged-in admin loses "active" status', async () => {
     const agent = supertest.agent(server)
     await agent